99% of bug on R_Knob is fixed

+++ src/rope/vector/vec2.java
@@ -689,7 +689,7 @@ public class vec2 extends vec {
 	 * @return float
 	 */
 	public float average() {
-		return (this.x()+this.y())*(float)0.5f;
+		return (this.x()+this.y())*0.5f;
 	}
